feat: add camera states constants
